+++ src/euphorie/client/publish.py
@@ -100,6 +100,7 @@ def CopyToClient(survey, preview=False):
     else:
         copy.id = surveygroup.id
     copy.title = surveygroup.title
+    copy.obsolete = surveygroup.obsolete
     copy.evaluation_algorithm = surveygroup.evaluation_algorithm
     copy.version = source.id
     copy.published = datetime.datetime.now()